requirements.txt
openai
flask
python-dotenv
flask_cors